# Fix GELF string escaping