Fix #11 

Perf improvement too
```
goos: darwin
goarch: amd64
pkg: github.com/mdouchement/logger
cpu: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
           │ testouille/old.txt │         testouille/new.txt          │
           │       sec/op       │   sec/op     vs base                │
SlogGELF-8          4.915µ ± 5%   3.310µ ± 3%  -32.65% (p=0.000 n=10)

           │ testouille/old.txt │         testouille/new.txt          │
           │        B/op        │     B/op      vs base               │
SlogGELF-8         2.448Ki ± 1%   2.329Ki ± 1%  -4.87% (p=0.000 n=10)

           │ testouille/old.txt │         testouille/new.txt         │
           │     allocs/op      │ allocs/op   vs base                │
SlogGELF-8           63.00 ± 0%   42.00 ± 0%  -33.33% (p=0.000 n=10)
```